refactor: Mapper 테스트를 RepositoryTest에서 분리하여 MapperTest로 이동
- ChannelApiRepositoryTest를 @DataJpaTest로 복원, findById 복합 키 테스트만 유지
- ChannelApiMapperTest 신규 생성: findByIdSellerId 테스트 2건 이동
